requirements.txt
numpy
onnx
onnxconverter-common
onnxruntime-gpu
opencv-python
torch
torchvision
pillow
